Made some changes in quadrature test renamed symmetric_SBP
