+++ CMakeLists.txt
@@ -111,3 +111,8 @@ STENSEAL_ADD_TEST(test_1d_upwind_laplace_matrix)
 ADD_EXECUTABLE(test_1d_compact_laplace_matrix tests/test_1d_compact_laplace_matrix.cc)
 TARGET_LINK_LIBRARIES(test_1d_compact_laplace_matrix ${LIBNAME})
 STENSEAL_ADD_TEST(test_1d_compact_laplace_matrix)
+
+# test for quadrature
+ADD_EXECUTABLE(test_1d_quadrature tests/test_1d_quadrature.cc)
+TARGET_LINK_LIBRARIES(test_1d_quadrature ${LIBNAME})
+STENSEAL_ADD_TEST(test_1d_quadrature)
